Moved item series into a subfolder.
